Create controllers
For now these are the basic operations.
Probably will need to change what is returned later
when creating the front-end, because some of them
should probably be seperate requests.

Will need to create DTO's (Data Transfer Object)
so that the gotten post, put and other request DTO
not contain unnecessary data.