MultiRangeSlider bug fix
When a max value slider was slid to the min value
slider it would give a wrong value like 2.03.01
which would make it default to a wrong value, which
would fuck up the slider.

Tried to sanatize it, round it, remove extra dots
and values after it but nothing worked expect adding
toFixed to the setMax value... which still causes
errors in the console, but I don't know how to fix
any other way.

At least this fix doesn't make the slider jump to an
undefined value.