Refactor NovelCard to use common layout.
Refactored NovelCards to use a common layout from
NovelCardLayout.

As a result of this, the design has become more unified
and repeating css has been deleted. All of the css
that pertains to Novel Cards has been centralized to
NovelCard.module.css file (except novel-summary
which is also used elsewhere).